+++ src/features/create-post/index.ts
@@ -1,0 +1,1 @@
+export { PostModal } from "./ui/PostModal";
