+++ src/widgets/sidebar/ui/Sidebar.tsx
@@ -1,9 +1,10 @@
 import { Button, Navbar, Space } from "@mantine/core";
 
+import { PostModal } from "@/features/create-post";
 import type { User } from "@/shared/entities";
 import { sideBarWidthBase } from "@/shared/lib/const";
 import { useModal } from "@/shared/model";
-import { AccountButton, LoginButton, MainLinks, PostModal } from "@/shared/ui";
+import { AccountButton, LoginButton, MainLinks } from "@/shared/ui";
 
 interface SidebarProps {
 	currentUser: User;
